added support for type and instance based rewrite steps
